Refactored tactics
