+++ CMakeLists.txt
@@ -56,7 +56,7 @@ set(MANUAL_SOURCES
         ${PROJECT_SOURCE_DIR}/src/manual/JoystickHandler.cpp
         )
 
-set(NEW_SKILLS_SOURCES
+set(SKILLS_SOURCES
         ${PROJECT_SOURCE_DIR}/src/stp/Skill.cpp
         ${PROJECT_SOURCE_DIR}/src/stp/skills/Kick.cpp
         ${PROJECT_SOURCE_DIR}/src/stp/skills/Chip.cpp
@@ -65,25 +65,26 @@ set(NEW_SKILLS_SOURCES
         ${PROJECT_SOURCE_DIR}/src/stp/skills/Shoot.cpp
         )
 
-set(NEW_TACTICS_SOURCES
+set(TACTICS_SOURCES
         ${PROJECT_SOURCE_DIR}/src/stp/Tactic.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/TestTactic.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/DriveWithBall.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/KickAtPos.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/GetBall.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/BlockRobot.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/Receive.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/ChipAtPos.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/ShootAtPos.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/Intercept.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/Halt.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/TestTactic.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/DriveWithBall.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/KickAtPos.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/GetBall.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/BlockRobot.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/Receive.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/ChipAtPos.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/ShootAtPos.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/Intercept.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/Halt.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/Formation.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/GetBallInDirection.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/BlockBall.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/KeeperBlockBall.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/AvoidBall.cpp
+        ${PROJECT_SOURCE_DIR}/src/stp/tactics/PositionAndAim.cpp
+
         ${PROJECT_SOURCE_DIR}/src/stp/new_constants/ControlConstants.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/Formation.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/GetBallInDirection.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/BlockBall.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/KeeperBlockBall.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/AvoidBall.cpp
-        ${PROJECT_SOURCE_DIR}/src/stp/new_tactics/PositionAndAim.cpp
         )
 
 set(NEW_ROLES_SOURCES
@@ -256,8 +257,8 @@ set(WORLD_SOURCES
 set(AI_SOURCES
         ${PROJECT_SOURCE_DIR}/src/ApplicationManager.cpp
         ${UTILS_SOURCES}
-        ${NEW_SKILLS_SOURCES}
-        ${NEW_TACTICS_SOURCES}
+        ${SKILLS_SOURCES}
+        ${TACTICS_SOURCES}
         ${NEW_ROLES_SOURCES}
         ${NEW_PLAYS_SOURCES}
         ${INVARIANTS_SOURCES}
