+++ src/stp/new_roles/Defender.cpp
@@ -2,8 +2,8 @@
 // Created by jordi on 26-03-20.
 //
 
-#include "stp/new_tactics/Intercept.h"
-#include "stp/new_tactics/BlockRobot.h"
+#include "stp/tactics/Intercept.h"
+#include "stp/tactics/BlockRobot.h"
 #include "stp/new_roles/Defender.h"
 
 namespace rtt::ai::stp::role {
