+++ src/stp/new_roles/TestRole.cpp
@@ -4,7 +4,7 @@
 
 #include <roboteam_utils/Print.h>
 #include <stp/new_roles/TestRole.h>
-#include <stp/new_tactics/TestTactic.h>
+#include <stp/tactics/TestTactic.h>
 
 namespace rtt::ai::stp {
