+++ src/stp/tactics/DriveWithBall.cpp
@@ -2,10 +2,10 @@
 // Created by timovdk on 3/16/20.
 //
 
-#include <roboteam_utils/Print.h>
-#include <stp/skills/GoToPos.h>
-#include <stp/skills/Rotate.h>
-#include <stp/new_tactics/DriveWithBall.h>
+#include "stp/tactics/DriveWithBall.h"
+
+#include "stp/skills/GoToPos.h"
+#include "stp/skills/Rotate.h"
 
 namespace rtt::ai::stp::tactic {
