+++ src/stp/tactics/Formation.cpp
@@ -2,7 +2,7 @@
 // Created by timovdk on 3/27/20.
 //
 
-#include "stp/new_tactics/Formation.h"
+#include "stp/tactics/Formation.h"
 
 #include "stp/skills/GoToPos.h"
 #include "stp/skills/Rotate.h"
