+++ src/stp/tactics/GetBall.cpp
@@ -2,14 +2,12 @@
 // Created by ratoone on 10-03-20.
 //
 
-#include "stp/new_tactics/GetBall.h"
-
-#include <utilities/GameStateManager.hpp>
+#include "stp/tactics/GetBall.h"
 
 #include "control/ControlUtils.h"
 #include "stp/skills/GoToPos.h"
 #include "stp/skills/Rotate.h"
-#include "world/FieldComputations.h"
+#include "utilities/GameStateManager.hpp"
 
 namespace rtt::ai::stp::tactic {
 GetBall::GetBall() { skills = collections::state_machine<Skill, Status, StpInfo>{skill::GoToPos(), skill::Rotate()}; }
@@ -21,6 +19,8 @@ std::optional<StpInfo> GetBall::calculateInfoForSkill(StpInfo const &info) noexc
 
     Vector2 robotPosition = info.getRobot().value()->getPos();
     Vector2 ballPosition = info.getBall().value()->getPos();
+
+    // If this robot is not the keeper, don't get the ball inside a defense area
     if (info.getRobot()->get()->getId() != GameStateManager::getCurrentGameState().keeperId && FieldComputations::pointIsInDefenseArea(info.getField().value(), ballPosition)) {
         ballPosition = control::ControlUtils::projectPositionToOutsideDefenseArea(info.getField().value(), ballPosition, control_constants::AVOID_BALL_DISTANCE);
     }
