+++ src/stp/tactics/Receive.cpp
@@ -2,7 +2,7 @@
 // Created by jordi on 13-03-20.
 //
 
-#include "stp/new_tactics/Receive.h"
+#include "stp/tactics/Receive.h"
 
 #include "stp/skills/GoToPos.h"
 #include "stp/skills/Rotate.h"
