+++ src/stp/tactics/TestTactic.cpp
@@ -2,10 +2,9 @@
 // Created by roboteam on 9/3/20.
 //
 
-#include "stp/new_tactics/TestTactic.h"
+#include "stp/tactics/TestTactic.h"
 
-#include <stp/skills/GoToPos.h>
-#include <stp/skills/Rotate.h>
+#include "stp/skills/GoToPos.h"
 
 namespace rtt::ai::stp {
 
@@ -16,30 +15,31 @@ TestTactic::TestTactic() {
 
 std::optional<StpInfo> TestTactic::calculateInfoForSkill(StpInfo const &info) noexcept {
     StpInfo skillStpInfo = info;
-    if(!skillStpInfo.getField()) return std::nullopt;
+    if (!skillStpInfo.getField()) return std::nullopt;
 
     auto length = skillStpInfo.getField()->getFieldLength();
     auto width = skillStpInfo.getField()->getFieldWidth();
 
+    // Set positions based on the current_num() of the state machine
+    // This switch will let a robot drive in a square
     switch (skills.current_num()) {
-    case 0:
-        skillStpInfo.setPositionToMoveTo(Vector2(length / 8, width / 8));
-        break;
-    case 1:
-        skillStpInfo.setPositionToMoveTo(Vector2(length / 8, -width / 8));
-        break;
-    case 2:
-        skillStpInfo.setPositionToMoveTo(Vector2(-length / 8, -width / 8));
-        break;
-    case 3:
-        skillStpInfo.setPositionToMoveTo(Vector2(-length / 8, width / 8));
-        break;
-    case 4:
-        skillStpInfo.setPositionToMoveTo(Vector2(length / 8, width / 8));
-        skills.reset();
-        break;
+        case 0:
+            skillStpInfo.setPositionToMoveTo(Vector2(length / 8, width / 8));
+            break;
+        case 1:
+            skillStpInfo.setPositionToMoveTo(Vector2(length / 8, -width / 8));
+            break;
+        case 2:
+            skillStpInfo.setPositionToMoveTo(Vector2(-length / 8, -width / 8));
+            break;
+        case 3:
+            skillStpInfo.setPositionToMoveTo(Vector2(-length / 8, width / 8));
+            break;
+        case 4:
+            skillStpInfo.setPositionToMoveTo(Vector2(length / 8, width / 8));
+            skills.reset();
+            break;
     }
-
 
     return skillStpInfo;
 }
@@ -53,8 +53,6 @@ bool TestTactic::isEndTactic() noexcept {
     return false;
 }
 
-const char *TestTactic::getName() {
-    return "Test Tactic";
-}
+const char *TestTactic::getName() { return "Test Tactic"; }
 
 }  // namespace rtt::ai::stp
